Extract shell value types
