+++ clients/apple/AlanNative/ShellModel.swift
@@ -1,214 +1,5 @@
 import Foundation
 
-enum ShellAttentionState: String, Codable, CaseIterable {
-    case idle
-    case active
-    case awaitingUser = "awaiting_user"
-    case notable
-}
-
-enum ShellTabKind: String, Codable, CaseIterable {
-    case terminal
-    case scratch
-    case log
-}
-
-enum ShellPaneTreeKind: String, Codable {
-    case split
-    case pane
-}
-
-enum ShellSplitDirection: String, Codable {
-    case horizontal
-    case vertical
-}
-
-enum ShellPaneSplitDirection: String, Codable, CaseIterable {
-    case left
-    case right
-    case up
-    case down
-
-    var splitDirection: ShellSplitDirection {
-        switch self {
-        case .left, .right:
-            return .vertical
-        case .up, .down:
-            return .horizontal
-        }
-    }
-
-    var placesNewPaneBeforeTarget: Bool {
-        switch self {
-        case .left, .up:
-            return true
-        case .right, .down:
-            return false
-        }
-    }
-
-    static func defaultPlacement(for splitDirection: ShellSplitDirection) -> ShellPaneSplitDirection {
-        switch splitDirection {
-        case .horizontal:
-            return .down
-        case .vertical:
-            return .right
-        }
-    }
-}
-
-enum ShellSpatialFocusDirection: String, Codable, CaseIterable {
-    case left
-    case right
-    case up
-    case down
-
-    var splitDirection: ShellSplitDirection {
-        switch self {
-        case .left, .right:
-            return .vertical
-        case .up, .down:
-            return .horizontal
-        }
-    }
-
-    var movesForward: Bool {
-        switch self {
-        case .right, .down:
-            return true
-        case .left, .up:
-            return false
-        }
-    }
-}
-
-enum ShellWorkspaceCommand: String, CaseIterable, Identifiable {
-    case newTerminalTab
-    case newAlanTab
-    case splitLeft
-    case splitRight
-    case splitUp
-    case splitDown
-    case focusLeft
-    case focusRight
-    case focusUp
-    case focusDown
-    case equalizeSplits
-    case closePane
-    case closeTab
-
-    var id: String { rawValue }
-}
-
-enum ShellLaunchTarget: String, Codable, CaseIterable {
-    case shell
-    case alan
-}
-
-struct ShellProcessBinding: Codable, Equatable {
-    let program: String
-    let argvPreview: [String]?
-
-    private enum CodingKeys: String, CodingKey {
-        case program
-        case argvPreview = "argv_preview"
-    }
-}
-
-struct ShellContextSnapshot: Codable, Equatable {
-    let workingDirectoryName: String?
-    let repositoryRoot: String?
-    let gitBranch: String?
-    let controlPath: String?
-    let socketPath: String?
-    let alanBindingFile: String?
-    let launchCommand: String?
-    let launchStrategy: String?
-    let shellIntegrationSource: String?
-    let processState: String?
-    let rendererPhase: String?
-    let rendererHealth: String?
-    let surfaceReadiness: String?
-    let inputReady: Bool?
-    let readonly: Bool?
-    let terminalMode: String?
-    let displayName: String?
-    let displayID: String?
-    let windowTitle: String?
-    let lastMetadataAt: String?
-    let lastCommandExitCode: Int?
-
-    init(
-        workingDirectoryName: String?,
-        repositoryRoot: String?,
-        gitBranch: String?,
-        controlPath: String?,
-        socketPath: String? = nil,
-        alanBindingFile: String?,
-        launchCommand: String? = nil,
-        launchStrategy: String?,
-        shellIntegrationSource: String?,
-        processState: String?,
-        rendererPhase: String? = nil,
-        rendererHealth: String? = nil,
-        surfaceReadiness: String? = nil,
-        inputReady: Bool? = nil,
-        readonly: Bool? = nil,
-        terminalMode: String? = nil,
-        displayName: String? = nil,
-        displayID: String? = nil,
-        windowTitle: String? = nil,
-        lastMetadataAt: String?,
-        lastCommandExitCode: Int?
-    ) {
-        self.workingDirectoryName = workingDirectoryName
-        self.repositoryRoot = repositoryRoot
-        self.gitBranch = gitBranch
-        self.controlPath = controlPath
-        self.socketPath = socketPath
-        self.alanBindingFile = alanBindingFile
-        self.launchCommand = launchCommand
-        self.launchStrategy = launchStrategy
-        self.shellIntegrationSource = shellIntegrationSource
-        self.processState = processState
-        self.rendererPhase = rendererPhase
-        self.rendererHealth = rendererHealth
-        self.surfaceReadiness = surfaceReadiness
-        self.inputReady = inputReady
-        self.readonly = readonly
-        self.terminalMode = terminalMode
-        self.displayName = displayName
-        self.displayID = displayID
-        self.windowTitle = windowTitle
-        self.lastMetadataAt = lastMetadataAt
-        self.lastCommandExitCode = lastCommandExitCode
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case workingDirectoryName = "working_directory_name"
-        case repositoryRoot = "repository_root"
-        case gitBranch = "git_branch"
-        case controlPath = "control_path"
-        case socketPath = "socket_path"
-        case alanBindingFile = "alan_binding_file"
-        case launchCommand = "launch_command"
-        case launchStrategy = "launch_strategy"
-        case shellIntegrationSource = "shell_integration_source"
-        case processState = "process_state"
-        case rendererPhase = "renderer_phase"
-        case rendererHealth = "renderer_health"
-        case surfaceReadiness = "surface_readiness"
-        case inputReady = "input_ready"
-        case readonly
-        case terminalMode = "terminal_mode"
-        case displayName = "display_name"
-        case displayID = "display_id"
-        case windowTitle = "window_title"
-        case lastMetadataAt = "last_metadata_at"
-        case lastCommandExitCode = "last_command_exit_code"
-    }
-}
-
 func shellUserFacingSummary(_ summary: String?) -> String? {
     guard let summary else { return nil }
