+++ clients/apple/scripts/test-shell-split-model.sh
@@ -10,6 +10,7 @@ TEST_BINARY="${BUILD_DIR}/shell-split-model-tests"
 mkdir -p "$MODULE_CACHE_DIR"
 
 CLANG_MODULE_CACHE_PATH="$MODULE_CACHE_DIR" swiftc \
+    "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellValueTypes.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/ShellModel.swift" \
     "$REPO_ROOT/clients/apple/scripts/test-shell-split-model.swift" \
     -o "$TEST_BINARY"
