+++ clients/apple/scripts/test-terminal-runtime-service.sh
@@ -10,6 +10,7 @@ TEST_BINARY="${BUILD_DIR}/terminal-runtime-service-tests"
 mkdir -p "$MODULE_CACHE_DIR"
 
 CLANG_MODULE_CACHE_PATH="$MODULE_CACHE_DIR" swiftc \
+    "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellValueTypes.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/ShellModel.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/ShellControlPlane.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/TerminalHostRuntime.swift" \
